middleware/MODBUS_SNMP/OLD/requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
easysnmp==0.2.5
getmac==0.8.3
idna==3.3
minimalmodbus==2.0.1
numpy==1.19.5
paho-mqtt==1.6.1
psutil==5.9.0
pycodestyle==2.8.0
pyModbusTCP==0.1.10
pyserial==3.5
python-dotenv==0.19.2
requests==2.27.1
toml==0.10.2
urllib3==1.26.8
